Add files via upload
The dataset consists of image data which is divided into two categories: Open Eyes and Closed Eyes. This sample dataset was taken from Kaggle for simplicity of use. The dataset was already cleaned and thus, was ready to be worked upon.

The dataset has been divided into training and testing datasets at a 75:25 ratio.

Using tensorflow and keras frameworks, the model is defined, compiled and summarized.